requirements.txt
flask
flask-pymongo
flask-jwt-extended
flask-cors
python-dotenv